refactor(racing-car): refactor codes
