+++ src/main/kotlin/racingcar/Main.kt
@@ -1,11 +1,13 @@
 package racingcar
 
+import racingcar.model.RacingCar
 import racingcar.utils.InputUtils
 
 fun main() {
     val inputView = InputUtils.askQuestions()
 
     val racingCar = RacingCar.ready(inputView)
+
     for (i in 1..inputView.gameCount) {
         val resultViews = racingCar.runOnce()
         resultViews.printResultS()
