+++ src/main/kotlin/racingcar/model/Car.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.model
 
 class Car(var score: Int = 0) {
     fun move(canMove: Boolean) {
