+++ src/main/kotlin/racingcar/model/MoveStrategy.kt
@@ -1,8 +1,9 @@
-package racingcar
+package racingcar.model
+
+private const val MIN_NUMBER_TO_MOVE = 4
 
 class MoveStrategy {
     companion object {
-        private const val MIN_NUMBER_TO_MOVE = 4
         fun canMove(randomNumber: Int): Boolean {
             return randomNumber >= MIN_NUMBER_TO_MOVE
         }
