+++ src/main/kotlin/racingcar/utils/RandomNumber.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.utils
 
 import java.util.Random
