+++ src/main/kotlin/racingcar/view/ResultView.kt
@@ -1,9 +1,11 @@
 package racingcar.view
 
+private const val ONE_BLOCK = "-"
+
 class ResultView(val score: Int) {
     fun printResult() {
         for (i in 1..score) {
-            print("-")
+            print(ONE_BLOCK)
         }
         println()
     }
