+++ src/main/kotlin/racingcar/view/ResultViews.kt
@@ -1,8 +1,10 @@
 package racingcar.view
 
+private const val LINE_SEPARATOR = "========================"
+
 class ResultViews(var resultViews: List<ResultView>) {
     fun printResultS() {
         resultViews.forEach { resultView: ResultView -> resultView.printResult() }
-        println("----------------------")
+        println(LINE_SEPARATOR)
     }
 }
