+++ src/test/kotlin/racingcar/CarTest.kt
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.model.Car
 
 internal class CarTest {
     @ParameterizedTest
