+++ src/test/kotlin/racingcar/CarsTest.kt
@@ -3,6 +3,7 @@ package racingcar
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.model.Cars
 
 internal class CarsTest {
     @ParameterizedTest
