+++ src/test/kotlin/racingcar/MoveStrategyTest.kt
@@ -3,6 +3,7 @@ package racingcar
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
+import racingcar.model.MoveStrategy
 
 internal class MoveStrategyTest {
     @ParameterizedTest
