+++ src/test/kotlin/racingcar/RandomNumberTest.kt
@@ -3,6 +3,7 @@ package racingcar
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.utils.RandomNumber
 
 internal class RandomNumberTest {
     @ParameterizedTest
